requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiosmtplib
email-validator
alembic